Split QueryLoader and QueryLoaderFile
